+++ exch/emsmdb/oxomsg.cpp
@@ -35,10 +35,6 @@ enum { /* for PR_SUBMIT_FLAGS (unused in Gromox) */
 	SUBMITFLAG_PREPROCESS = 0x2U,
 };
 
-enum class repr_grant {
-	error = -1, no_impersonation, send_on_behalf, send_as,
-};
-
 /**
  * @send_as:	if true, copy PR_SENT_REPR to PR_SENDER
  * 		if false, leave PR_SENT_REPR at its value
@@ -424,8 +420,7 @@ ec_error_t rop_submitmessage(uint8_t submit_flags, LOGMAP *plogmap,
 		return ecSuccess;
 	}
 
-	auto ev_from = repr_grant >= repr_grant::send_as ? delegator.c_str() : actor;
-	ret = cu_send_message(plogon, pmessage, ev_from);
+	ret = cu_send_message(plogon, pmessage, actor, delegator.c_str(), repr_grant);
 	if (ret != ecSuccess && ret != ecWarnWithErrors)
 		exmdb_client->clear_submit(dir, pmessage->get_id(), b_unsent);
 	else if (!b_delete)
@@ -661,8 +656,7 @@ ec_error_t rop_transportsend(TPROPVAL_ARRAY **pppropvals, LOGMAP *plogmap,
 		}
 	}
 
-	auto ev_from = repr_grant >= repr_grant::send_as ? delegator.c_str() : actor;
-	return cu_send_message(plogon, pmessage, ev_from);
+	return cu_send_message(plogon, pmessage, actor, delegator.c_str(), repr_grant);
 } catch (const std::bad_alloc &) {
 	mlog(LV_ERR, "%s: ENOMEM", __func__);
 	return ecServerOOM;